+++ src/Sylius/Bundle/ApiBundle/spec/StateProcessor/Admin/Country/PersistProcessorSpec.php
@@ -11,9 +11,8 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Bundle\ApiBundle\StateProcessor;
+namespace spec\Sylius\Bundle\ApiBundle\StateProcessor\Admin\Country;
 
-use ApiPlatform\Metadata\Delete;
 use ApiPlatform\Metadata\HttpOperation;
 use ApiPlatform\Metadata\Post;
 use ApiPlatform\State\ProcessorInterface;
@@ -23,26 +22,23 @@ use Sylius\Bundle\ApiBundle\Exception\ProvinceCannotBeRemoved;
 use Sylius\Component\Addressing\Checker\CountryProvincesDeletionCheckerInterface;
 use Sylius\Component\Addressing\Model\CountryInterface;
 
-final class CountryProcessorSpec extends ObjectBehavior
+final class PersistProcessorSpec extends ObjectBehavior
 {
     function let(
         ProcessorInterface $persistProcessor,
-        ProcessorInterface $removeProcessor,
         CountryProvincesDeletionCheckerInterface $countryProvincesDeletionChecker,
     ): void {
-        $this->beConstructedWith($persistProcessor, $removeProcessor, $countryProvincesDeletionChecker);
+        $this->beConstructedWith($persistProcessor, $countryProvincesDeletionChecker);
     }
 
     function it_throws_an_exception_if_object_is_not_a_country(
         ProcessorInterface $persistProcessor,
-        ProcessorInterface $removeProcessor,
         CountryProvincesDeletionCheckerInterface $countryProvincesDeletionChecker,
         HttpOperation $operation,
     ): void {
         $countryProvincesDeletionChecker->isDeletable(Argument::any())->shouldNotBeCalled();
 
         $persistProcessor->process(Argument::cetera())->shouldNotBeCalled();
-        $removeProcessor->process(Argument::cetera())->shouldNotBeCalled();
 
         $this
             ->shouldThrow(\InvalidArgumentException::class)
@@ -50,27 +46,8 @@ final class CountryProcessorSpec extends ObjectBehavior
         ;
     }
 
-    function it_uses_decorated_data_persister_to_remove_country(
-        ProcessorInterface $persistProcessor,
-        ProcessorInterface $removeProcessor,
-        CountryProvincesDeletionCheckerInterface $countryProvincesDeletionChecker,
-        CountryInterface $country,
-    ): void {
-        $operation = new Delete();
-        $uriVariables = [];
-        $context = [];
-
-        $countryProvincesDeletionChecker->isDeletable($country)->shouldNotBeCalled();
-
-        $persistProcessor->process(Argument::cetera())->shouldNotBeCalled();
-        $removeProcessor->process($country, $operation, $uriVariables, $context)->willReturn($country);
-
-        $this->process($country, $operation, $uriVariables, $context)->shouldReturn($country);
-    }
-
     function it_uses_decorated_data_persister_to_persist_country(
         ProcessorInterface $persistProcessor,
-        ProcessorInterface $removeProcessor,
         CountryProvincesDeletionCheckerInterface $countryProvincesDeletionChecker,
         CountryInterface $country,
     ): void {
@@ -80,7 +57,6 @@ final class CountryProcessorSpec extends ObjectBehavior
 
         $countryProvincesDeletionChecker->isDeletable($country)->willReturn(true);
 
-        $removeProcessor->process(Argument::cetera())->shouldNotBeCalled();
         $persistProcessor->process($country, $operation, $uriVariables, $context)->willReturn($country);
 
         $this->process($country, $operation, $uriVariables, $context)->shouldReturn($country);
@@ -88,7 +64,6 @@ final class CountryProcessorSpec extends ObjectBehavior
 
     function it_throws_an_error_if_the_province_within_a_country_is_in_use(
         ProcessorInterface $persistProcessor,
-        ProcessorInterface $removeProcessor,
         CountryProvincesDeletionCheckerInterface $countryProvincesDeletionChecker,
         CountryInterface $country,
         HttpOperation $operation,
@@ -99,7 +74,6 @@ final class CountryProcessorSpec extends ObjectBehavior
         $countryProvincesDeletionChecker->isDeletable($country)->willReturn(false);
 
         $persistProcessor->process(Argument::cetera())->shouldNotBeCalled();
-        $removeProcessor->process(Argument::cetera())->shouldNotBeCalled();
 
         $this
             ->shouldThrow(ProvinceCannotBeRemoved::class)
